refactor: domain 모듈로 단위 테스트 이동
