+++ domain/build.gradle.kts
@@ -7,3 +7,6 @@ java {
     sourceCompatibility = JavaVersion.VERSION_1_7
     targetCompatibility = JavaVersion.VERSION_1_7
 }
+dependencies {
+    implementation("junit:junit:4.12")
+}
