+++ domain/src/test/java/MorningPolicyTest.kt
@@ -1,5 +1,3 @@
-package woowacourse.movie
-
 import com.example.domain.model.policy.MorningPolicy
 import com.example.domain.model.price.Price
 import junit.framework.TestCase.assertEquals
@@ -7,13 +5,12 @@ import org.junit.Test
 import java.time.LocalDate
 import java.time.LocalTime
 
-class NightPolicyTest {
-
+class MorningPolicyTest {
     @Test
-    fun `오후 8시 이후 상영 시간은 2천원 할인된다`() {
+    fun `오전 11시 이전 상영 시간은 2천원 할인된다`() {
         val actual = MorningPolicy()
             .calculate(
-                LocalDate.of(2023, 4, 11), LocalTime.of(8, 0),
+                LocalDate.of(2023, 4, 11), LocalTime.of(9, 0),
                 Price()
             )
         val expected = Price(11000)
