+++ domain/src/test/java/MovieDayPolicyTest.kt
@@ -1,5 +1,3 @@
-package woowacourse.movie
-
 import com.example.domain.model.price.Price
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
