+++ domain/src/test/java/PlayingTimesTest.kt
@@ -1,5 +1,3 @@
-package woowacourse.movie
-
 import com.example.domain.model.model.PlayingTimes
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
